remove pyontutils deps
A note on isoformat. While in principle we should try to
keep isoformat in sync with what is used in pyontutils,
the key thing here is that the representation remain
consisten when converting PathMeta objects to other
formats. Thus, keeping all the relevant implementation in
meta.py seems like the best decision.