RepoPath.rmtree fixes for windows
RepoPath.commit hits a code path inside GitPython that makes a call to
git.cmd.Git._get_persistent_cmd which spawns processes that have their
working dir inside a repository, so if we try to rmtree the repository
windows gets unhappy and yells at us that some program is still using
the path. That someone is ofcourse our own subprocesses, so we pop off
the offending repo class and close it.

The two offending calls inside of commit are to repo.active_branch.is_valid()
the is_valid is the primary offender, not the active_branch, and
repo.git.stash is the other. There may be others, but this commit
should fix all of them. Appropriately, if you try to rmtree using a
different type of path then the error will still occure on windows as
it should since only RepoPath classes should be rmtreeing themselves
if there are active repo instances in a running python program.

Given that git.Repo is probably supposed to be used inside a context
manager and ours are not for performance/object duplication reasons,
it might be a good idea to see if there are some consistent points at
which it makes sense to pop repos off in general and close them.