repo defer importing git until required
Many use use cases for augpathlib do not involve RepoPath so we defer
importing git until the first time a RepoPath is instantiated. I'm sure
that there is some tradeoff that will come back to haunt me by doing
this, but forcing anyone who wants to use a library to pay the worst
case time cost for the maximal use case seems at least as bad.