LocalPath fix issue with alternate data streams and 'wb' on ntfs
So the semantics of 'wb' on windows and posix are different with
respect to ADS and xattrs respectively. On posix 'wb' preserves xattrs
because it doesn't modify the inode, on windows I have no idea what it
does, but the end result is that any extant data streams are removed,
which was causing no end of confusion, and seems like a fundamental
failing in the specification and behavior of open.

Many more tests are needed to check for this issue.